src/test/java/org/zubov/solutions/easy/LongestCommonPrefixTest.java
package org.zubov.solutions.easy;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.junit.jupiter.api.Assertions.*;

class LongestCommonPrefixTest {

    private static Stream<Arguments> provideParameters() {
        return Stream.of(
                Arguments.of(new String[]{"flower", "flow", "flight"}, "fl"),
                Arguments.of(new String[]{"dog", "racecar", "car"}, "")
        );
    }

    @ParameterizedTest
    @MethodSource("provideParameters")
    void testCases(String[] words, String expected) {
        //Given

        //When
        String result = LongestCommonPrefix.longestCommonPrefix(words);

        //Then
        assertEquals(expected, result);
    }
}